+++ effect_subject/tests/tests/test_signs_and_symptoms.py
@@ -37,30 +37,30 @@ class TestFollowupFormValidation(EffectTestCaseMixin, TestCase):
         super().setUp()
         self.subject_visit = self.get_subject_visit()
 
-    # TODO: Test any_signs_symptoms == Unknown
+    # TODO: Test any_sx == Unknown
     # TODO: Test patient with signs and symptoms is valid
 
     def get_valid_patient_with_no_signs_or_symptoms(self):
         return {
             "appointment": self.subject_visit.appointment,
             "report_datetime": self.subject_visit.report_datetime,
-            "any_signs_symptoms": NO,
-            "signs_and_symptoms": SiSx.objects.none(),
-            "signs_and_symptoms_gte_g3": SiSx.objects.none(),
+            "any_sx": NO,
+            "current_sx": SiSx.objects.none(),
+            "current_sx_gte_g3": SiSx.objects.none(),
             "headache_duration": "",
             "visual_field_loss": "",
             "reportable_as_ae": NOT_APPLICABLE,
             "patient_admitted": NOT_APPLICABLE,
-            "cm_signs_symptoms": NOT_APPLICABLE,
+            "cm_sx": NOT_APPLICABLE,
         }
 
-    def get_valid_patient_any_signs_or_symptoms_unknown(self):
+    def get_valid_patient_any_sx_unknown(self):
         cleaned_data = deepcopy(self.get_valid_patient_with_no_signs_or_symptoms())
         cleaned_data.update(
             {
-                "any_signs_symptoms": UNKNOWN,
-                "signs_and_symptoms": SiSx.objects.none(),
-                "signs_and_symptoms_gte_g3": SiSx.objects.none(),
+                "any_sx": UNKNOWN,
+                "current_sx": SiSx.objects.none(),
+                "current_sx_gte_g3": SiSx.objects.none(),
             }
         )
         return cleaned_data
@@ -69,23 +69,21 @@ class TestFollowupFormValidation(EffectTestCaseMixin, TestCase):
         return {
             "appointment": self.subject_visit.appointment,
             "report_datetime": self.subject_visit.report_datetime,
-            "any_signs_symptoms": YES,
-            "signs_and_symptoms": SiSx.objects.filter(
-                Q(name="fever") | Q(name="vomiting")
-            ),
-            "signs_and_symptoms_gte_g3": SiSx.objects.none(),
+            "any_sx": YES,
+            "current_sx": SiSx.objects.filter(Q(name="fever") | Q(name="vomiting")),
+            "current_sx_gte_g3": SiSx.objects.none(),
             "headache_duration": "",
             "visual_field_loss": "",
             "reportable_as_ae": NO,
             "patient_admitted": NO,
-            "cm_signs_symptoms": NO,
+            "cm_sx": NO,
         }
 
     def get_valid_patient_with_g3_signs_or_symptoms(self):
         cleaned_data = deepcopy(self.get_valid_patient_with_signs_or_symptoms())
         cleaned_data.update(
             {
-                "signs_and_symptoms_gte_g3": SiSx.objects.filter(name="fever"),
+                "current_sx_gte_g3": SiSx.objects.filter(name="fever"),
                 "reportable_as_ae": YES,
             }
         )
@@ -97,8 +95,8 @@ class TestFollowupFormValidation(EffectTestCaseMixin, TestCase):
             form_validator=self.validate_form_validator(cleaned_data)
         )
 
-    def test_unknown_any_sas(self):
-        cleaned_data = self.get_valid_patient_any_signs_or_symptoms_unknown()
+    def test_any_sx_unknown_valid(self):
+        cleaned_data = self.get_valid_patient_any_sx_unknown()
         self.assertFormValidatorNoError(
             form_validator=self.validate_form_validator(cleaned_data)
         )
@@ -115,45 +113,45 @@ class TestFollowupFormValidation(EffectTestCaseMixin, TestCase):
             form_validator=self.validate_form_validator(cleaned_data)
         )
 
-    def test_m2m_signs_and_symptoms_not_required_if_any_sas_is_not_yes(self):
-        any_sas_answers = [ans for ans, _ in YES_NO_UNKNOWN if ans != YES]
-        for answer in any_sas_answers:
+    def test_m2m_current_sx_not_required_if_any_sx_is_not_yes(self):
+        any_sx_answers = [ans for ans, _ in YES_NO_UNKNOWN if ans != YES]
+        for answer in any_sx_answers:
             with self.subTest(answer=answer):
-                cleaned_data = self.get_valid_patient_any_signs_or_symptoms_unknown()
+                cleaned_data = self.get_valid_patient_any_sx_unknown()
                 cleaned_data.update(
                     {
-                        "any_signs_symptoms": answer,
-                        "signs_and_symptoms": SiSx.objects.filter(name="fever"),
+                        "any_sx": answer,
+                        "current_sx": SiSx.objects.filter(name="fever"),
                     }
                 )
                 self.assertFormValidatorError(
-                    field="signs_and_symptoms",
+                    field="current_sx",
                     expected_msg="This field is not required",
                     form_validator=self.validate_form_validator(cleaned_data),
                 )
 
-        cleaned_data.update({"signs_and_symptoms": SiSx.objects.none()})
+        cleaned_data.update({"current_sx": SiSx.objects.none()})
         self.assertFormValidatorNoError(self.validate_form_validator(cleaned_data))
 
-    def test_m2m_signs_and_symptoms_required_if_any_sas_is_yes(self):
-        cleaned_data = self.get_valid_patient_any_signs_or_symptoms_unknown()
+    def test_m2m_current_sx_required_if_any_sx_is_yes(self):
+        cleaned_data = self.get_valid_patient_any_sx_unknown()
         cleaned_data.update(
             {
-                "any_signs_symptoms": YES,
-                "signs_and_symptoms": SiSx.objects.none(),
+                "any_sx": YES,
+                "current_sx": SiSx.objects.none(),
             }
         )
         self.assertFormValidatorError(
-            field="signs_and_symptoms",
+            field="current_sx",
             expected_msg="This field is required",
             form_validator=self.validate_form_validator(cleaned_data),
         )
 
-    def test_reportable_as_ae_applicable_if_any_signs_symptoms_yes(self):
+    def test_reportable_as_ae_applicable_if_any_sx_yes(self):
         cleaned_data = self.get_valid_patient_with_signs_or_symptoms()
         cleaned_data.update(
             {
-                "any_signs_symptoms": YES,
+                "any_sx": YES,
                 "reportable_as_ae": NOT_APPLICABLE,
             }
         )
@@ -167,7 +165,7 @@ class TestFollowupFormValidation(EffectTestCaseMixin, TestCase):
             with self.subTest(answer=answer):
                 cleaned_data.update(
                     {
-                        "signs_and_symptoms_gte_g3": SiSx.objects.filter(name="fever")
+                        "current_sx_gte_g3": SiSx.objects.filter(name="fever")
                         if answer == YES
                         else SiSx.objects.none(),
                         "reportable_as_ae": answer,
@@ -177,18 +175,18 @@ class TestFollowupFormValidation(EffectTestCaseMixin, TestCase):
                     form_validator=self.validate_form_validator(cleaned_data)
                 )
 
-    def test_reportable_as_ae_not_applicable_if_any_signs_symptoms_not_yes(self):
-        for any_sas_answer in [NO, UNKNOWN]:
+    def test_reportable_as_ae_not_applicable_if_any_sx_not_yes(self):
+        for any_sx_answer in [NO, UNKNOWN]:
             for reportable_as_ae_answer in [YES, NO]:
                 with self.subTest(
-                    any_sas_answer=any_sas_answer,
+                    any_sx_answer=any_sx_answer,
                     reportable_as_ae_answer=reportable_as_ae_answer,
                 ):
                     cleaned_data = self.get_valid_patient_with_no_signs_or_symptoms()
                     cleaned_data.update(
                         {
-                            "any_signs_symptoms": any_sas_answer,
-                            "signs_and_symptoms": SiSx.objects.none(),
+                            "any_sx": any_sx_answer,
+                            "current_sx": SiSx.objects.none(),
                             "reportable_as_ae": reportable_as_ae_answer,
                         }
                     )
@@ -203,62 +201,60 @@ class TestFollowupFormValidation(EffectTestCaseMixin, TestCase):
                         form_validator=self.validate_form_validator(cleaned_data)
                     )
 
-    def test_signs_and_symptoms_gte_g3_not_required_if_reportable_as_ae_is_not_yes(
+    def test_current_sx_gte_g3_not_required_if_reportable_as_ae_is_not_yes(
         self,
     ):
         for reportable_as_ae_answer in [NO, NOT_APPLICABLE]:
             with self.subTest(reportable_as_ae_answer=reportable_as_ae_answer):
                 cleaned_data = self.get_valid_patient_with_no_signs_or_symptoms()
                 cleaned_data.update(
-                    {"signs_and_symptoms_gte_g3": SiSx.objects.filter(name="fever")}
+                    {"current_sx_gte_g3": SiSx.objects.filter(name="fever")}
                 )
 
                 self.assertFormValidatorError(
-                    field="signs_and_symptoms_gte_g3",
+                    field="current_sx_gte_g3",
                     expected_msg="This field is not required",
                     form_validator=self.validate_form_validator(cleaned_data),
                 )
 
                 cleaned_data = self.get_valid_patient_with_no_signs_or_symptoms()
-                cleaned_data.update({"signs_and_symptoms_gte_g3": SiSx.objects.none()})
+                cleaned_data.update({"current_sx_gte_g3": SiSx.objects.none()})
                 self.assertFormValidatorNoError(
                     form_validator=self.validate_form_validator(cleaned_data)
                 )
 
-    def test_signs_and_symptoms_gte_g3_required_if_reportable_as_ae_is_yes(self):
+    def test_current_sx_gte_g3_required_if_reportable_as_ae_is_yes(self):
         cleaned_data = self.get_valid_patient_with_g3_signs_or_symptoms()
         cleaned_data.update(
             {
-                "signs_and_symptoms_gte_g3": SiSx.objects.none(),
+                "current_sx_gte_g3": SiSx.objects.none(),
                 "reportable_as_as": YES,
             }
         )
         self.assertFormValidatorError(
-            field="signs_and_symptoms_gte_g3",
+            field="current_sx_gte_g3",
             expected_msg="This field is required",
             form_validator=self.validate_form_validator(cleaned_data),
         )
 
-        cleaned_data.update(
-            {"signs_and_symptoms_gte_g3": SiSx.objects.filter(name="fever")}
-        )
+        cleaned_data.update({"current_sx_gte_g3": SiSx.objects.filter(name="fever")})
         self.assertFormValidatorNoError(
             form_validator=self.validate_form_validator(cleaned_data)
         )
 
-    def test_signs_and_symptoms_gte_g3_not_required_if_any_sas_is_yes(self):
+    def test_current_sx_gte_g3_not_required_if_any_sx_is_yes(self):
         cleaned_data = self.get_valid_patient_with_signs_or_symptoms()
-        cleaned_data.update({"signs_and_symptoms_gte_g3": SiSx.objects.none()})
+        cleaned_data.update({"current_sx_gte_g3": SiSx.objects.none()})
 
         self.assertFormValidatorNoError(
             form_validator=self.validate_form_validator(cleaned_data),
         )
 
-    def test_headache_duration_applicable_if_in_signs_and_symptoms(self):
+    def test_headache_duration_applicable_if_in_current_sx(self):
         cleaned_data = self.get_valid_patient_with_signs_or_symptoms()
         cleaned_data.update(
             {
-                "signs_and_symptoms": SiSx.objects.filter(name=HEADACHE),
+                "current_sx": SiSx.objects.filter(name=HEADACHE),
                 "headache_duration": "",
             }
         )
@@ -273,7 +269,7 @@ class TestFollowupFormValidation(EffectTestCaseMixin, TestCase):
             form_validator=self.validate_form_validator(cleaned_data)
         )
 
-    def test_headache_duration_not_applicable_if_not_in_signs_and_symptoms(self):
+    def test_headache_duration_not_applicable_if_not_in_current_sx(self):
         cleaned_data = self.get_valid_patient_with_no_signs_or_symptoms()
         cleaned_data.update({"headache_duration": "2d3h"})
         self.assertFormValidatorError(
@@ -298,11 +294,11 @@ class TestFollowupFormValidation(EffectTestCaseMixin, TestCase):
             form_validator=self.validate_form_validator(cleaned_data)
         )
 
-    def test_visual_field_loss_applicable_if_in_signs_and_symptoms(self):
+    def test_visual_field_loss_applicable_if_in_current_sx(self):
         cleaned_data = self.get_valid_patient_with_signs_or_symptoms()
         cleaned_data.update(
             {
-                "signs_and_symptoms": SiSx.objects.filter(name=VISUAL_LOSS),
+                "current_sx": SiSx.objects.filter(name=VISUAL_LOSS),
                 "visual_field_loss": "",
             }
         )
@@ -317,7 +313,7 @@ class TestFollowupFormValidation(EffectTestCaseMixin, TestCase):
             form_validator=self.validate_form_validator(cleaned_data)
         )
 
-    def test_visual_field_loss_not_applicable_if_not_in_signs_and_symptoms(self):
+    def test_visual_field_loss_not_applicable_if_not_in_current_sx(self):
         cleaned_data = self.get_valid_patient_with_no_signs_or_symptoms()
         cleaned_data.update({"visual_field_loss": "Details on visual field loss"})
         self.assertFormValidatorError(
@@ -342,12 +338,12 @@ class TestFollowupFormValidation(EffectTestCaseMixin, TestCase):
             form_validator=self.validate_form_validator(cleaned_data)
         )
 
-    def test_patient_admitted_applicable_if_any_signs_symptoms_yes(self):
+    def test_patient_admitted_applicable_if_any_sx_yes(self):
         cleaned_data = self.get_valid_patient_with_signs_or_symptoms()
         cleaned_data.update(
             {
-                "any_signs_symptoms": YES,
-                "signs_and_symptoms": SiSx.objects.filter(name="fever"),
+                "any_sx": YES,
+                "current_sx": SiSx.objects.filter(name="fever"),
                 "patient_admitted": NOT_APPLICABLE,
             }
         )
@@ -364,18 +360,18 @@ class TestFollowupFormValidation(EffectTestCaseMixin, TestCase):
                     form_validator=self.validate_form_validator(cleaned_data)
                 )
 
-    def test_patient_admitted_not_applicable_if_any_signs_symptoms_not_yes(self):
-        for any_sas_answer in [NO, UNKNOWN]:
+    def test_patient_admitted_not_applicable_if_any_sx_not_yes(self):
+        for any_sx_answer in [NO, UNKNOWN]:
             for patient_admitted_answer in [YES, NO]:
                 with self.subTest(
-                    any_sas_answer=any_sas_answer,
+                    any_sx_answer=any_sx_answer,
                     patient_admitted_answer=patient_admitted_answer,
                 ):
                     cleaned_data = self.get_valid_patient_with_no_signs_or_symptoms()
                     cleaned_data.update(
                         {
-                            "any_signs_symptoms": any_sas_answer,
-                            "signs_and_symptoms": SiSx.objects.none(),
+                            "any_sx": any_sx_answer,
+                            "current_sx": SiSx.objects.none(),
                             "patient_admitted": patient_admitted_answer,
                         }
                     )
@@ -390,53 +386,51 @@ class TestFollowupFormValidation(EffectTestCaseMixin, TestCase):
                         form_validator=self.validate_form_validator(cleaned_data)
                     )
 
-    def test_cm_signs_symptoms_applicable_if_any_signs_symptoms_yes(self):
+    def test_cm_sx_applicable_if_any_sx_yes(self):
         cleaned_data = self.get_valid_patient_with_signs_or_symptoms()
         cleaned_data.update(
             {
-                "any_signs_symptoms": YES,
-                "signs_and_symptoms": SiSx.objects.filter(
-                    Q(name="fever") | Q(name="vomiting")
-                ),
+                "any_sx": YES,
+                "current_sx": SiSx.objects.filter(Q(name="fever") | Q(name="vomiting")),
                 "patient_admitted": NO,
-                "cm_signs_symptoms": NOT_APPLICABLE,
+                "cm_sx": NOT_APPLICABLE,
             }
         )
         self.assertFormValidatorError(
-            field="cm_signs_symptoms",
+            field="cm_sx",
             expected_msg="This field is applicable.",
             form_validator=self.validate_form_validator(cleaned_data),
         )
 
         for answer in [YES, NO]:
             with self.subTest(answer=answer):
-                cleaned_data.update({"cm_signs_symptoms": answer})
+                cleaned_data.update({"cm_sx": answer})
                 self.assertFormValidatorNoError(
                     form_validator=self.validate_form_validator(cleaned_data)
                 )
 
-    def test_cm_signs_symptoms_not_applicable_if_any_signs_symptoms_not_yes(self):
-        for any_sas_answer in [NO, UNKNOWN]:
-            for cm_signs_symptoms_answer in [YES, NO]:
+    def test_cm_sx_not_applicable_if_any_sx_not_yes(self):
+        for any_sx_answer in [NO, UNKNOWN]:
+            for cm_sx_answer in [YES, NO]:
                 with self.subTest(
-                    any_sas_answer=any_sas_answer,
-                    cm_signs_symptoms_answer=cm_signs_symptoms_answer,
+                    any_sx_answer=any_sx_answer,
+                    cm_sx_answer=cm_sx_answer,
                 ):
                     cleaned_data = self.get_valid_patient_with_no_signs_or_symptoms()
                     cleaned_data.update(
                         {
-                            "any_signs_symptoms": any_sas_answer,
-                            "signs_and_symptoms": SiSx.objects.none(),
-                            "cm_signs_symptoms": cm_signs_symptoms_answer,
+                            "any_sx": any_sx_answer,
+                            "current_sx": SiSx.objects.none(),
+                            "cm_sx": cm_sx_answer,
                         }
                     )
                     self.assertFormValidatorError(
-                        field="cm_signs_symptoms",
+                        field="cm_sx",
                         expected_msg="This field is not applicable.",
                         form_validator=self.validate_form_validator(cleaned_data),
                     )
 
-                    cleaned_data.update({"cm_signs_symptoms": NOT_APPLICABLE})
+                    cleaned_data.update({"cm_sx": NOT_APPLICABLE})
                     self.assertFormValidatorNoError(
                         form_validator=self.validate_form_validator(cleaned_data)
                     )